:bug:Fix: delivery __str__ return fix, product avg star round #9 #10
