--- products/tests/__init__.py
@@ -1,15 +1,0 @@
-from .product_tests import (
-    BaseTestCase,
-    ProductCreateTest,
-    ProductListTest,
-    ProductDetailTest,
-    ProductUpdateTest,
-    ProductDeleteTest,
-)
-from .review_tests import (
-    ReviewCreateTest,
-    ReviewListTest,
-    ReviewDetailTest,
-    ReviewUpdateTest,
-    ReviewDeleteTest,
-)
